# Improve FragmentTombstones() speed by lazily initializing `seq_set_`